pyproject.toml
[tool.poetry]
name = "mountain-peak-backend"
version = "0.1.0"
description = "MFI recrutment process. Simple web service for storing and retrieving moutain peaks."
authors = ["Simon Lyobard <simon.lyobard@gmail.com>"]
readme = "README.md"
packages = [{include = "mountain_peak_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.1.3"
djangorestframework = "^3.14.0"
django-filter = "^22.1"
django-extensions = "^3.2.1"
pydantic = "^1.10.2"
Markdown = "^3.4.1"
psycopg2-binary = "^2.9.5"
ipython = "^8.6.0"
python-dotenv = "^0.21.0"
drf-extra-fields = "^3.4.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
black = "^22.10.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 150

[tool.isort]
profile = "black"